requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cloudinary==1.41.0
dj-database-url==0.5.0
Django==3.2
django-cors-headers==4.4.0
django-heroku==0.3.1
djangorestframework==3.12.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
